More convenient API for backup
