+++ Tests/SQLiteTests/ConnectionTests.swift
@@ -141,6 +141,18 @@ class ConnectionTests : SQLiteTestCase {
 
         AssertSQL("BEGIN EXCLUSIVE TRANSACTION")
     }
+    
+    func test_backup_copiesDatabase() throws {
+        let target = try Connection()
+        
+        try InsertUsers("alice", "betsy")
+        
+        let backup = try db.backup(usingConnection: target)
+        try backup.step()
+        
+        let users = try target.prepare("SELECT email FROM users ORDER BY email")
+        XCTAssertEqual(users.map { $0[0] as? String }, ["alice@example.com", "betsy@example.com"])
+    }
 
     func test_transaction_beginsAndCommitsTransactions() {
         let stmt = try! db.prepare("INSERT INTO users (email) VALUES (?)", "alice@example.com")
